Add error handler on controller. New get user tracks endompoint

+++ src/server/controllers/Track.controller.js
@@ -2,27 +2,44 @@ import * as fs from 'node:fs/promises';
 import path from 'node:path';
 import { Track } from '../models/models.js';
 
-// todo error handling
-
 const uploadTrack = async (req, res) => {
-  const { username } = req.params;
-  const { originalname, filename, size } = req.file;
-  const newTrack = new Track({
-    uploaded_by: username,
-    path: filename,
-    title: originalname,
-    size: size,
-    date: new Date(),
-  });
-  await newTrack.save();
-  res.status(200);
-  res.send(newTrack);
+  try {
+    const { username } = req.params;
+    const { originalname, filename, size } = req.file;
+    const newTrack = new Track({
+      uploaded_by: username,
+      path: filename,
+      title: originalname,
+      size: size,
+      date: new Date(),
+    });
+    await newTrack.save();
+    res.status(200).send(newTrack);
+  } catch (error) {
+    console.log(error);
+    res.status(500).send({ error });
+  }
 };
 
 const getAllTracks = async (req, res) => {
-  const tracks = await Track.find({});
-  res.status(200);
-  res.send(tracks);
+  try {
+    const tracks = await Track.find({});
+    res.status(200).send(tracks);
+  } catch (error) {
+    console.log(error);
+    res.status(500).send({ error });
+  }
+};
+
+const getUserTracks = async (req, res) => {
+  try {
+    const { username } = req.params;
+    const tracks = await Track.find({ user: username });
+    res.status(200).send(tracks);
+  } catch (error) {
+    console.log(error);
+    res.status(500).send({ error });
+  }
 };
 
 const tracksPublicDirectory = './public/tracks'; // path relative to the node process
@@ -32,14 +49,14 @@ const deleteTrack = async (req, res) => {
     // The name of the file is the id of the track, and the path at the same time.
     const { id } = req.params;
     const { deletedCount } = await Track.deleteOne({ path: id });
-    await fs.unlink(path.join(tracksPublicDirectory, id));
-    res.status(200);
-    res.send({ deletedCount });
+    if (deletedCount) {
+      await fs.unlink(path.join(tracksPublicDirectory, id));
+    }
+    res.status(200).send({ deletedCount });
   } catch (error) {
     console.log(error);
-    res.status(500);
-    res.send({ error });
+    res.status(500).send({ error });
   }
 };
 
-export { uploadTrack, getAllTracks, deleteTrack };
+export { uploadTrack, getAllTracks, deleteTrack, getUserTracks };
